Add 2 classes (Astroid sys) and frame handler
The classes have been added to allow for astroid to move and astroid sys to spawn and soooo on. (Look at the classes for a bit more into it)

THE IMPORTANT PART

have added a handler for the app to all functions to be called every frame. (its calls every frame)

THE INPORTENT PART

have added a handler for the app to all functions to be called every frame. (its calls every frame)

P.S Sorry for any spelling mistakes in the comments :)